Reorganized tests #34
