+++ backend/tests/test_attendance.py
@@ -1,0 +1,8 @@
+"""Tests for attendance functionality."""
+from django.test import TestCase
+from .test_utils import BaseTestCase
+
+
+class AttendanceTestCase(BaseTestCase):
+    """Tests for attendance endpoints (placeholder)."""
+    pass
