+++ backend/tests/test_prices.py
@@ -1,0 +1,16 @@
+"""Tests for price management functionality."""
+import json
+from django.test import TestCase
+from .test_utils import BaseTestCase
+
+
+class PricesTestCase(BaseTestCase):
+    """Tests for price endpoints (GET /prices/, POST /prices/)."""
+
+    def setUp(self):
+        super().setUp()
+        self.prices_url = "/api/prices/"
+
+    # TODO: Add tests for price creation, retrieval, update, delete
+    # TODO: Test price filtering by class, date ranges, etc.
+    pass
